requirements-vercel.txt
# Vercel-compatible requirements for SET Portfolio App

# Web Framework and API
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
pydantic==2.5.0
numpy==1.24.3

# Database (Supabase)
supabase==2.0.2

# Web Scraping and HTTP
requests==2.31.0
beautifulsoup4==4.12.2
httpx==0.25.2
urllib3==2.1.0

# Excel File Processing
openpyxl==3.1.2
xlrd==2.0.1

# Environment Variables
python-dotenv==1.0.0

# Financial Data (for mini charts)
yfinance==0.2.28
lxml==4.9.3

# SSL and Security
certifi==2023.11.17
cryptography==41.0.8

# Additional dependencies for better compatibility
aiofiles==23.2.1
python-dateutil==2.8.2

# Note: playwright is removed for Vercel deployment as it requires system dependencies
# Browser automation features will be disabled in production
